+++ src/widget/button.rs
@@ -83,7 +83,7 @@ impl<'a, Message> Button<'a, Message> {
     /// Sets the width of the [`Button`] in pixels.
     ///
     /// [`Button`]: struct.Button.html
-    pub fn width(mut self, width: u32) -> Self {
+    pub fn width(mut self, width: u16) -> Self {
         self.style = self.style.width(width);
         self
     }
