+++ src/widget/row.rs
@@ -52,7 +52,7 @@ impl<'a, Message, Renderer> Row<'a, Message, Renderer> {
     /// Sets the padding of the [`Row`] in pixels.
     ///
     /// [`Row`]: struct.Row.html
-    pub fn padding(mut self, px: u32) -> Self {
+    pub fn padding(mut self, px: u16) -> Self {
         self.style = self.style.padding(px);
         self
     }
@@ -60,7 +60,7 @@ impl<'a, Message, Renderer> Row<'a, Message, Renderer> {
     /// Sets the width of the [`Row`] in pixels.
     ///
     /// [`Row`]: struct.Row.html
-    pub fn width(mut self, width: u32) -> Self {
+    pub fn width(mut self, width: u16) -> Self {
         self.style = self.style.width(width);
         self
     }
@@ -68,7 +68,7 @@ impl<'a, Message, Renderer> Row<'a, Message, Renderer> {
     /// Sets the height of the [`Row`] in pixels.
     ///
     /// [`Row`]: struct.Row.html
-    pub fn height(mut self, height: u32) -> Self {
+    pub fn height(mut self, height: u16) -> Self {
         self.style = self.style.height(height);
         self
     }
@@ -76,7 +76,7 @@ impl<'a, Message, Renderer> Row<'a, Message, Renderer> {
     /// Sets the maximum width of the [`Row`] in pixels.
     ///
     /// [`Row`]: struct.Row.html
-    pub fn max_width(mut self, max_width: u32) -> Self {
+    pub fn max_width(mut self, max_width: u16) -> Self {
         self.style = self.style.max_width(max_width);
         self
     }
@@ -84,7 +84,7 @@ impl<'a, Message, Renderer> Row<'a, Message, Renderer> {
     /// Sets the maximum height of the [`Row`] in pixels.
     ///
     /// [`Row`]: struct.Row.html
-    pub fn max_height(mut self, max_height: u32) -> Self {
+    pub fn max_height(mut self, max_height: u16) -> Self {
         self.style = self.style.max_height(max_height);
         self
     }
